pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]

[project]
name = "ghostwriter-pro-hotstream"
version = "0.1.0"
description = "多平台数据抓取框架"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "playwright>=1.40.0",
    "asyncio",
    "aiofiles>=23.2.1",
    "pydantic>=2.5.0",
    "python-crontab>=3.0.0",
    "sqlalchemy>=2.0.23",
    "pymongo>=4.6.0",
    "redis>=5.0.1",
    "requests>=2.31.0",
    "aiohttp>=3.9.1",
    "beautifulsoup4>=4.12.2",
    "lxml>=4.9.3",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "uvloop>=0.19.0",
    "psutil>=5.9.6",
    "schedule>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
]

[project.scripts]
hotstream = "hotstream.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["hotstream*"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88